+++ SmartStudent/src/main/java/com/manage/admin/dao/Admin_BacklogDAOImpl.java
@@ -32,7 +32,7 @@ public class Admin_BacklogDAOImpl implements Admin_BacklogDAO {
     }
 
     @Override
-    public void deleteBacklog(Integer studentId, String subjectCode, Integer semester) {
+    public void deleteBacklog(Long studentId, String subjectCode, Integer semester) {
         Backlog.IdClass id = new Backlog.IdClass(studentId, subjectCode, semester);
         Backlog backlog = sessionFactory.getCurrentSession().get(Backlog.class, id);
         if (backlog != null) {
@@ -41,7 +41,7 @@ public class Admin_BacklogDAOImpl implements Admin_BacklogDAO {
     }
 
     @Override
-    public Backlog getBacklogById(Integer studentId, String subjectCode, Integer semester) {
+    public Backlog getBacklogById(Long studentId, String subjectCode, Integer semester) {
         Backlog.IdClass id = new Backlog.IdClass(studentId, subjectCode, semester);
         return sessionFactory.getCurrentSession().get(Backlog.class, id);
     }
